requirements.txt
numpy==2.2.5
opencv_python==4.11.0.86
Pillow==11.2.1
streamlit==1.45.0
tensorflow==2.19.0
